Move bundle configuration filename code
