+++ bundle/config/root.go
@@ -12,36 +12,6 @@ import (
 	"github.com/imdario/mergo"
 )
 
-type ConfigFileNames []string
-
-// FileNames contains allowed names of bundle configuration files.
-var FileNames = ConfigFileNames{"databricks.yml", "databricks.yaml", "bundle.yml", "bundle.yaml"}
-
-func (c ConfigFileNames) FindInPath(path string) (string, error) {
-	result := ""
-	var firstErr error
-
-	for _, file := range c {
-		filePath := filepath.Join(path, file)
-		_, err := os.Stat(filePath)
-		if err == nil {
-			if result != "" {
-				return "", fmt.Errorf("multiple bundle root configuration files found in %s", path)
-			}
-			result = filePath
-		} else {
-			if firstErr == nil {
-				firstErr = err
-			}
-		}
-	}
-
-	if result == "" {
-		return "", firstErr
-	}
-	return result, nil
-}
-
 type Root struct {
 	// Path contains the directory path to the root of the bundle.
 	// It is set when loading `databricks.yml`.
